+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixes
 
 - Prevents exception capture context from being overwritten by native scope sync ([#4124](https://github.com/getsentry/sentry-react-native/pull/4124))
+- Skips development server spans ([#4271](https://github.com/getsentry/sentry-react-native/pull/4271))
 
 ## 6.2.0
 
@@ -39,7 +40,6 @@
 
 ### Fixes
 
-- Skips development server spans ([#4271](https://github.com/getsentry/sentry-react-native/pull/4271))
 - Ignore JavascriptException to filter out obfuscated duplicate JS Errors on Android ([#4232](https://github.com/getsentry/sentry-react-native/pull/4232))
 - Skips ignoring require cycle logs for RN 0.70 or newer ([#4214](https://github.com/getsentry/sentry-react-native/pull/4214))
 - Enhanced accuracy of time-to-display spans. ([#4189](https://github.com/getsentry/sentry-react-native/pull/4189))